Very performant react-virtualized tree (for collections view)
Unlike the mozilla's version, this one is extremely fast even with
thousands of rows and does not display any row pop-in while scrolling.
Moreover, going up and down by holding down [Page Up], or just arrow
keys works very smoothly.

The update includes a multi-select functionality from the get-go, as
well as a flattened row structure for the input, which makes the API
for this tree much more compatible with existing code used with
nsITreeVIew